+++ oak-jcr/src/test/java/org/apache/jackrabbit/oak/jcr/OakSegmentTarRepositoryStub.java
@@ -24,6 +24,7 @@ import javax.jcr.Repository;
 import javax.jcr.RepositoryException;
 
 import org.apache.jackrabbit.oak.Oak;
+import org.apache.jackrabbit.oak.plugins.document.repoinit.BundlingConfigInitializer;
 import org.apache.jackrabbit.oak.segment.SegmentNodeStoreBuilders;
 import org.apache.jackrabbit.oak.segment.file.FileStore;
 import org.apache.jackrabbit.oak.segment.file.FileStoreBuilder;
@@ -71,7 +72,7 @@ public class OakSegmentTarRepositoryStub extends BaseRepositoryStub {
      * {@link Jcr} builder before the repository is created. This default
      * implementation set query engine settings as returned by
      * {@link #getQueryEngineSettings()} and adds a
-     * {@link org.apache.jackrabbit.oak.plugins.document.bundlor.BundlingConfigInitializer}.
+     * {@link BundlingConfigInitializer}.
      *
      * @param jcr the builder.
      * @param whiteboard the oak whiteboard
