+++ oak-store-document/src/main/java/org/apache/jackrabbit/oak/plugins/document/repoinit/BundlingConfigInitializer.java
@@ -17,9 +17,10 @@
  * under the License.
  */
 
-package org.apache.jackrabbit.oak.plugins.document.bundlor;
+package org.apache.jackrabbit.oak.plugins.document.repoinit;
 
 import org.apache.jackrabbit.oak.api.Type;
+import org.apache.jackrabbit.oak.plugins.document.bundlor.BundledTypesRegistry;
 import org.apache.jackrabbit.oak.spi.nodetype.NodeTypeConstants;
 import org.apache.jackrabbit.oak.spi.lifecycle.RepositoryInitializer;
 import org.apache.jackrabbit.oak.spi.state.NodeBuilder;
